+++ src/util.rs
@@ -6,6 +6,7 @@ use proc_macro2::{Ident, Span, TokenStream};
 use quote::quote;
 use std::collections::HashSet;
 use std::path::{Path, PathBuf};
+use svd_parser::expand::BlockPath;
 use svd_rs::{MaybeArray, PeripheralInfo};
 
 use syn::{
@@ -439,6 +440,12 @@ pub fn path_segment(ident: Ident) -> PathSegment {
         ident,
         arguments: PathArguments::None,
     }
+}
+
+pub fn parent(p: &BlockPath) -> BlockPath {
+    let mut p = p.clone();
+    p.path.pop().unwrap();
+    p
 }
 
 pub trait U32Ext {